Generalized ListSymbols() to SearchSymbols() and use it in ListSymbols() and SymbolName();
Summary:
SearchSymbols() will be used to find a function's address, offset in binary file, and size.
Then we can extract the function's byte code from the .text section.
And finally to extract the return instructions.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-1067

Differential Revision: https://phab.corp.pixielabs.ai/D4176

GitOrigin-RevId: cb4ec1888bc3ac2df0a9a640aad9e6a41456b41f

+++ src/common/base/byte_utils.h
@@ -39,7 +39,7 @@ T ReverseBytes(T x) {
  * @param buf The sequence of bytes.
  * @return The decoded int value.
  */
-template <typename T, int N = sizeof(T)>
+template <typename T, size_t N = sizeof(T)>
 T LEndianBytesToInt(std::string_view buf) {
   // Doesn't make sense to process more bytes than the destination type.
   // Less bytes is okay, on the other hand, since the value will still fit.
